Calculate activation
The network is activated by propagating signals through it until
each output node is active. The outputs will be active all the
time after the start phase. The signals are sent through the
network in timesteps. So each node sends its output to each
node one step away, which does the same the next step and so on.